requirements.txt
flask==2.2.2
kafka-python==2.0.2
pandas==1.5.3
matplotlib==3.7.1
openpyxl==3.0.10
